requirements.txt
PyYAML>=6.0
# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0